+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   but a simple int year*10000 + month*100 + day.  Should make creation of Dates 
   faster which is useful when parsing json for example. 
 
+## 2026-08-28
+- Add method years() to term, to return the list of years that cover the term.
+
 ## 2026-07-24
 - Add method months() to term, to return the list of months that cover the term.